+++ app/helpers/katas_helper.rb
@@ -22,14 +22,14 @@ module KatasHelper
 
   def delete_external_kata_button(id)
     text = t(:delete)
-    return link_to(text, external_kata_path(id), { method: :delete, class: "btn btn-danger"}) if admin_signed_in?
+    return link_to(text, external_kata_path(id), { method: :delete, class: "btn btn-danger", data: {confirm: "Are you sure?"}}) if admin_signed_in?
 
     NO_LINK
   end
 
   def delete_belt_button(id)
     text = t(:delete_belt)
-    return link_to(text, delete_belt_path(id), { method: :delete, class: "btn btn-danger"}) if admin_signed_in?
+    return link_to(text, delete_belt_path(id), { method: :delete, class: "btn btn-danger", data: {confirm: "Are you sure?"}}) if admin_signed_in?
 
     NO_LINK
   end
@@ -57,7 +57,7 @@ module KatasHelper
 
   def delete_kata_button(kata)
     text = t(:delete)
-    return link_to(text, kata_path(kata), { method: :delete, class: "btn btn-danger"}) if admin_signed_in?
+    return link_to(text, kata_path(kata), { method: :delete, class: "btn btn-danger", data: {confirm: "Are you sure?"}}) if admin_signed_in?
 
     NO_LINK
   end
